Put in a check to notify if have already found the record

+++ discogs.sh
@@ -1,6 +1,15 @@
 #!/bin/bash
 
-lynx -cmd_script=./record_script1 -accept_all_cookies http://www.google.co.uk/search?q=discogs+$1
+exists=`find . -name $1.txt`
 
-grep -A 10 Statistics 1 > $1.txt
-rm 1
+if ["$exists" eq ""]
+then
+	lynx -cmd_script=./record_script -accept_all_cookies http://www.google.co.uk/search?q=discogs+$1
+
+	grep -A 10 Statistics 1 > Outputs/$1.txt
+	rm 1
+
+else
+	echo "Search exists"
+
+fi
